refactor: allow unused locals
This is a temporary hack to allow unused imports. When the type gen runs
it always write the following line:

import { SoapMappingBase, propsWithMappingsName } from "../../soap-types";

In the case of transactions_financials.ts, SoapMappingBase is unused.
Rather than change the type generator to handle this (the correct fix),
this allows for unused locals.